split ResourceMonitor to a diff file
